# api-docs/routes/openapi.yaml
openapi: 3.0.1
info:
  title: UZH Masters Project API Documentation
  description: "The following list of API endpoints connected by the API Gateway to the different Microservices. Click [Try it out] to test the endpoints and edit the Example Value."
  contact:
    email: calvinmitchell.ongkingco@uzh.ch
  version: 1.14.3

# Gateway Link
servers:
- url: https://uzhmp-api-gateway-77xdzfzvua-ew.a.run.app/api/
  description: The gateway link of all endpoint services

# Tags for the endpoint services.
...
paths:
...
  /v2/user/list:
    get:
      tags:
      - user
      summary: (Admin Access) Shows list of Users
      description: (Admin Access) Gets the list of all the users
      operationId: getUsers
      responses:
        "200":
          description: Success
        "401":
          description: Unauthorized
        "404":
          description: User Record Does Not Exist
        "500":
          description: Error in Searching User Profile
      security:
      - tokenAuth: []
      x-swagger-router-controller: User
  /v2/user/profile:
    get:
      tags:
      - user
      summary: "Current user details"
      description: "Current user's details"
      operationId: "getUserProfileDetails"
      responses:
        "200":
          description: "User details"
        "403":
          description: "Unauthorized"
        "500":
          description: "Error"
      security:
        - tokenAuth: []
    put:
      tags:
      - user
      summary: "(Admin Access) Update a User Profile - (Will follow up on a different commit)"
      description: "(Admin Access) Update a User Profile"
      servers:
      - url: https://
      operationId: updateUserProfile
      responses:
        "200":
          description: Successfully Updated Item
        "404":
          description: No User Record Found
        "500":
          description: Error in Updating User Profile
      deprecated: true
      security:
      - tokenAuth: []
      x-swagger-router-controller: User

# IMAGE ENDPOINTS
  /imsrv:
    get:
      tags:
      - image
      summary: Lists all available uploaded images
      servers:
      - url: https://uzhmp-api-gateway-77xdzfzvua-ew.a.run.app/api/v1/apg
      description: Lists all available images that have been uploaded (Admin sees the rest)
        and available in the database
      operationId: getImageList
      responses:
        "200":
          description: (list of images)
        "403":
          description: Invalid Token Provided or Token has already Expired.
        "500":
          description: Error in Searching Image Lists
      security:
      - tokenAuth: []
      x-swagger-router-controller: Image
  /imsrv/res/{id}:
    get:
      tags:
      - image
      summary: Get details of one specific image
      servers:
      - url: https://uzhmp-api-gateway-77xdzfzvua-ew.a.run.app/api/v1/apg
      description: Get details of an image
      operationId: getImageDetails
      parameters:
      - name: id
        in: path
        description: ID of image to get details
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: byte
      responses:
        "200":
          description: Successfully Uploaded Image
        "400":
          description: No File Uploaded.
        "403":
          description: Invalid Token Provided or Token has already Expired.
      security:
      - tokenAuth: []
      x-swagger-router-controller: Image
...
  /imsrv/res:
    put:
      tags:
      - image
      summary: Delete an image
      servers:
      - url: https://uzhmp-api-gateway-77xdzfzvua-ew.a.run.app/api/v1/apg
      description: Deletes an image using the ID
      operationId: deleteImageRecord
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteImage'
      responses:
        "202":
          description: Successfully Deleted Image
          content:
            application/json:
              example:
                message: Successfully Deleted Image
        "401":
          description: Invalid Token Provided or Token has already Expired.
      security:
      - tokenAuth: []
      x-swagger-router-controller: Image

# MAPPING SERVICES ENDPOINTS
...
components:
  schemas:
...
    DeleteImage:
      type: object
      properties:
        imageid:
          type: string
          description: The ID of the image
          example: PASTE THE ID HERE
      additionalProperties: false
...
  securitySchemes:
    tokenAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT